+++ flash/tabular/__init__.py
@@ -1,1 +1,2 @@
-from flash.tabular.classification import TabularClassifier, TabularData  # noqa: F401
+from flash.tabular.classification import TabularClassifier, TabularClassificationData  # noqa: F401
+from flash.tabular.data import TabularData  # noqa: F401
