+++ flash/tabular/classification/__init__.py
@@ -1,2 +1,2 @@
-from flash.tabular.classification.data import TabularData  # noqa: F401
+from flash.tabular.classification.data import TabularClassificationData  # noqa: F401
 from flash.tabular.classification.model import TabularClassifier  # noqa: F401
