+++ flash/tabular/regression/__init__.py
@@ -1,0 +1,1 @@
+from flash.tabular.regression.data import TabularRegressionData  # noqa: F401
